Fix fmt + clippy warnings across 28 files
- cargo fmt --all formatting pass
- Fix should_implement_trait: implement FromStr for AgentMode
- Fix collapsible_if, unnecessary_map_or, unused_enumerate_index
- Fix cloned_ref_to_slice_refs, useless_format, let_and_return
- Fix unnecessary_cast, useless_vec, redundant closures
- Fix unused variable, ptr_arg, type_complexity warnings

+++ tests/bfcl_pipeline.rs
@@ -248,16 +248,8 @@ async fn test_bfcl_voltr_pipeline() {
 
             let (pass, pt, ct) = match result {
                 Ok(r) => {
-                    let pt = r
-                        .usage
-                        .as_ref()
-                        .map(|u| u.prompt_tokens as u64)
-                        .unwrap_or(0);
-                    let ct = r
-                        .usage
-                        .as_ref()
-                        .map(|u| u.completion_tokens as u64)
-                        .unwrap_or(0);
+                    let pt = r.usage.as_ref().map(|u| u.prompt_tokens).unwrap_or(0);
+                    let ct = r.usage.as_ref().map(|u| u.completion_tokens).unwrap_or(0);
                     let ok = if let Some(ref tcs) = r.tool_calls {
                         let pred: Vec<&str> = tcs.iter().map(|tc| tc.name.as_str()).collect();
                         let exp: Vec<&str> = functions
